Move cgi and test more

+++ tests/cgi_test.py
@@ -8,7 +8,7 @@ import sys
 import time
 
 
-def run_cgi(cgibin, UNIXSOCK, envi, close_after_accept):
+def run_cgi(cgibin, UNIXSOCK, envi, close_after_accept, maxclient=1):
     ret = {}
     if os.path.exists(UNIXSOCK):
         os.unlink(UNIXSOCK)
@@ -24,9 +24,15 @@ def run_cgi(cgibin, UNIXSOCK, envi, close_after_accept):
     fcntl.fcntl(pp.stdout, fcntl.F_SETFL, flags)
     timeout = 0
     theend = False
-    while timeout < 10 and not theend:
+    numclient = 0
+    MAXTIMEOUT = 15
+    while timeout < MAXTIMEOUT and not theend:
         try:
-            c, addr = us.accept()
+            c, adidr = us.accept()
+            numclient += 1
+            if numclient == 2:
+                # ugly hack
+                close_after_accept = False
             if close_after_accept:
                 c.close()
             else:
@@ -34,13 +40,19 @@ def run_cgi(cgibin, UNIXSOCK, envi, close_after_accept):
                 ret["recv"] = r
                 c.send(b"SENDSTRING")
                 c.close()
-            theend = True
+            if numclient >= maxclient:
+                theend = True
         except BlockingIOError:
+            print(f"TIMEOUT {timeout}")
             timeout += 1
-        time.sleep(1)
+        time.sleep(0.3)
         pp.poll()
+    if timeout >= MAXTIMEOUT:
+        ret["error"] = "timeout"
     pp.stdout.flush()
     us.close()
+    if os.path.exists(UNIXSOCK):
+        os.remove(UNIXSOCK)
     outs, err = pp.communicate()
     print(f"OUT={outs}")
     ret["out"] = outs
@@ -57,7 +69,7 @@ def test_proxy():
 
     envi = {}
     envi["REQUEST_METHOD"] = 'invalid'
-    ret = subprocess.run(cgibin, capture_output=True, env=envi)
+    #ret = subprocess.run(cgibin, capture_output=True, env=envi)
     ret = subprocess.run(cgibin, capture_output=True, env=envi)
     print(ret)
     assert ret.returncode == 0
@@ -122,7 +134,7 @@ def test_proxy():
     assert ret.stdout == b'Content-type: text/plain\n\nERROR: no REMOTE_ADDR\n'
     assert ret.stderr == b''
 
-    UNIXSOCK = "tests/cgi/xython.sock"
+    UNIXSOCK = "tests/cgi/xython-proxy.sock"
     envi["XYTHON_SOCK"] = UNIXSOCK
 
     print("================================= FileNotFoundError:")
@@ -190,7 +202,8 @@ def test_proxy():
     assert ret["out"] == b'Content-type: text/plain\n\nSENDSTRING\n'
 
     us.close()
-    os.remove(UNIXSOCK)
+    if os.path.exists(UNIXSOCK):
+        os.remove(UNIXSOCK)
 
 
 def test_topchanges():
@@ -202,7 +215,7 @@ def test_topchanges():
 
     envi = {}
     envi["REQUEST_METHOD"] = 'invalid'
-    ret = subprocess.run(cgibin, capture_output=True, env=envi)
+    #ret = subprocess.run(cgibin, capture_output=True, env=envi)
     ret = subprocess.run(cgibin, capture_output=True, env=envi)
     print(ret)
     assert ret.returncode == 0
@@ -213,8 +226,8 @@ def test_topchanges():
     envi["REQUEST_METHOD"] = 'GET'
     ret = subprocess.run(cgibin, capture_output=True, env=envi)
     print(ret)
-    assert ret.returncode == 0
-    assert ret.stdout == b'Content-type: text/html\n\nERROR: no starttime\n'
+    assert ret.returncode == 1
+    assert ret.stdout == b'Content-type: text/html\n\nERROR: not runned as CGI\n'
     assert ret.stderr == b''
 
     envi["QUERY_STRING"] = 'invalid'
@@ -252,6 +265,13 @@ def test_topchanges():
     assert ret.stdout == b'Content-type: text/html\n\nERROR: invalid TOTIME date\n'
     assert ret.stderr == b''
 
+    envi["QUERY_STRING"] = 'FROMTIME=2023%2F01%2F01%4011%3A11%3A11'
+    ret = subprocess.run(cgibin, capture_output=True, env=envi)
+    print(ret)
+    assert ret.returncode == 0
+    assert ret.stdout == b'Content-type: text/html\n\nERROR: no TOTIME\n'
+    assert ret.stderr == b''
+
     envi["QUERY_STRING"] = 'FROMTIME=2023%2F01%2F01%4011%3A11%3A11&TOTIME=invalid'
     ret = subprocess.run(cgibin, capture_output=True, env=envi)
     print(ret)
@@ -259,7 +279,7 @@ def test_topchanges():
     assert ret.stdout == b'Content-type: text/html\n\nERROR: invalid TOTIME date\n'
     assert ret.stderr == b''
 
-    UNIXSOCK = "tests/cgi/xython.sock"
+    UNIXSOCK = "tests/cgi/xython-topchanges.sock"
     envi["XYTHON_SOCK"] = UNIXSOCK
 
     if os.path.exists(UNIXSOCK):
@@ -268,14 +288,14 @@ def test_topchanges():
     ret = subprocess.run(cgibin, capture_output=True, env=envi)
     print(ret)
     assert ret.returncode == 0
-    assert ret.stdout == b'Content-type: text/html\n\nFAIL to connect to xythond, no such file or directory\n'
+    assert ret.stdout == b'Content-type: text/html\n\nFAIL to connect to xythond, [Errno 2] No such file or directory\n'
     assert ret.stderr == b''
 
     envi["QUERY_STRING"] = 'FROMTIME=2023%2F01%2F01%4011%3A11%3A11&TOTIME=2023%2F01%2F01%4011%3A11%3A11'
     ret = subprocess.run(cgibin, capture_output=True, env=envi)
     print(ret)
     assert ret.returncode == 0
-    assert ret.stdout == b'Content-type: text/html\n\nFAIL to connect to xythond, no such file or directory\n'
+    assert ret.stdout == b'Content-type: text/html\n\nFAIL to connect to xythond, [Errno 2] No such file or directory\n'
     assert ret.stderr == b''
 
     us = socket.socket(socket.AF_UNIX, socket.SOCK_STREAM)
@@ -285,34 +305,12 @@ def test_topchanges():
     ret = subprocess.run(cgibin, capture_output=True, env=envi)
     print(ret)
     assert ret.returncode == 0
-    assert ret.stdout == b'Content-type: text/html\n\nFAIL to connect to xythond\n'
+    assert ret.stdout == b'Content-type: text/html\n\nFAIL to connect to xythond, [Errno 111] Connection refused\n'
     assert ret.stderr == b''
 
-    us.listen(10)
-    us.setblocking(0)
-
-    pp = subprocess.Popen('./cgi/topchanges.py', stdout=subprocess.PIPE, env=envi)
-    flags = fcntl.fcntl(pp.stdout, fcntl.F_GETFL)
-    flags = flags | os.O_NONBLOCK
-    fcntl.fcntl(pp.stdout, fcntl.F_SETFL, flags)
-    timeout = 0
-    theend = False
-    while timeout < 10 and not theend:
-        try:
-            c, addr = us.accept()
-            print("ACCEPT")
-            c.close()
-            theend = True
-        except BlockingIOError:
-            print("WAIT")
-            timeout += 1
-        time.sleep(1)
-        ret = pp.poll()
-    pp.stdout.flush()
     us.close()
-    outs, err = pp.communicate()
-    print(f"OUT={outs}")
-    print(f"ERR={err}")
+    if os.path.exists(UNIXSOCK):
+        os.remove(UNIXSOCK)
 
     ret = run_cgi(cgibin, UNIXSOCK, envi, True)
     print(f"RET={ret}")
@@ -324,4 +322,166 @@ def test_topchanges():
     assert ret["out"] == b'Content-type: text/html\n\nSENDSTRING\n'
 
     us.close()
-    os.remove(UNIXSOCK)
+    if os.path.exists(UNIXSOCK):
+        os.remove(UNIXSOCK)
+
+def test_getpages():
+    cgibin = [sys.executable, "-m", "coverage", 'run', './cgi/getpage.py']
+    ret = subprocess.run(cgibin, capture_output=True, env=None)
+    assert ret.returncode == 1
+    assert ret.stdout == b'Content-type: text/html\n\nERROR: not runned as CGI\n'
+    assert ret.stderr == b''
+
+    envi = {}
+    UNIXSOCK = "tests/cgi/xython.sock"
+    envi["XYTHON_SOCK"] = UNIXSOCK
+
+    if os.path.exists(UNIXSOCK):
+        os.unlink(UNIXSOCK)
+
+    ret = run_cgi(cgibin, UNIXSOCK, envi, True)
+    print(f"RET={ret}")
+    assert "out" in ret
+    assert ret["out"] == b'Content-type: text/html\n\nERROR: not runned as CGI\n'
+
+    envi["QUERY_STRING"] = 'page=toto'
+    ret = run_cgi(cgibin, UNIXSOCK, envi, False)
+    print(f"RET={ret}")
+    assert "out" in ret
+    assert ret["out"] == b'Content-type: text/html\n\nSENDSTRING\n'
+
+def test_xythoncgi():
+    cgibin = [sys.executable, "-m", "coverage", 'run', './cgi/xythoncgi.py']
+    ret = subprocess.run(cgibin, capture_output=True, env=None)
+    assert ret.returncode == 1
+    assert ret.stdout == b'Content-type: text/html\n\nERROR: not runned as CGI\n'
+    assert ret.stderr == b''
+
+    envi = {}
+    UNIXSOCK = "tests/cgi/xython-xythoncgi.sock"
+    envi["XYTHON_SOCK"] = UNIXSOCK
+
+    if os.path.exists(UNIXSOCK):
+        os.unlink(UNIXSOCK)
+
+    envi["QUERY_STRING"] = 'hostname=toto&service=test'
+    ret = subprocess.run(cgibin, capture_output=True, env=envi)
+    print(ret)
+    assert ret.returncode == 0
+    assert ret.stdout == b'Content-type: text/html\n\nFAIL to connect to xythond, [Errno 2] No such file or directory\n'
+    assert ret.stderr == b''
+
+    del(envi["QUERY_STRING"])
+    ret = run_cgi(cgibin, UNIXSOCK, envi, True)
+    print(f"RET={ret}")
+    assert "out" in ret
+    assert ret["out"] == b'Content-type: text/html\n\nERROR: not runned as CGI\n'
+
+    envi["QUERY_STRING"] = 'service=toto'
+    ret = run_cgi(cgibin, UNIXSOCK, envi, False)
+    print(f"RET={ret}")
+    assert "out" in ret
+    assert ret["out"] == b'Content-type: text/html\n\nERROR: no hostname\n'
+
+    envi["QUERY_STRING"] = 'hostname=toto'
+    ret = run_cgi(cgibin, UNIXSOCK, envi, False)
+    print(f"RET={ret}")
+    assert "out" in ret
+    assert ret["out"] == b'Content-type: text/html\n\nERROR: no service\n'
+
+    envi["QUERY_STRING"] = 'HOST=toto'
+    ret = run_cgi(cgibin, UNIXSOCK, envi, False)
+    print(f"RET={ret}")
+    assert "out" in ret
+    assert ret["out"] == b'Content-type: text/html\n\nERROR: no service\n'
+
+    envi["QUERY_STRING"] = 'hostname=toto&SERVICE=test'
+    ret = run_cgi(cgibin, UNIXSOCK, envi, False)
+    print(f"RET={ret}")
+    assert "out" in ret
+    assert ret["out"] == b'Content-type: text/html\n\nSENDSTRING\n'
+
+    envi["QUERY_STRING"] = 'HOST=toto&SERVICE=test'
+    ret = run_cgi(cgibin, UNIXSOCK, envi, False)
+    print(f"RET={ret}")
+    assert "out" in ret
+    assert ret["out"] == b'Content-type: text/html\n\nSENDSTRING\n'
+
+    envi["QUERY_STRING"] = 'hostname=toto&SERVICE=test&timebuf=toto'
+    ret = run_cgi(cgibin, UNIXSOCK, envi, False)
+    print(f"RET={ret}")
+    assert "out" in ret
+    assert ret["out"] == b'Content-type: text/html\n\nSENDSTRING\n'
+
+    envi["QUERY_STRING"] = 'hostname=toto&SERVICE=test&action=invalid'
+    ret = run_cgi(cgibin, UNIXSOCK, envi, False)
+    print(f"RET={ret}")
+    assert "out" in ret
+    assert ret["out"] == b'Content-type: text/html\n\nERROR: invalid action\n'
+
+    envi["QUERY_STRING"] = 'hostname=toto&SERVICE=test&action=ack'
+    ret = run_cgi(cgibin, UNIXSOCK, envi, False)
+    print(f"RET={ret}")
+    assert "out" in ret
+    assert ret["out"] == b'Content-type: text/html\n\nERROR: ack need cause\n'
+
+    envi["QUERY_STRING"] = 'hostname=toto&SERVICE=test&action=ack&cause=test'
+    ret = run_cgi(cgibin, UNIXSOCK, envi, False)
+    print(f"RET={ret}")
+    assert "out" in ret
+    assert ret["out"] == b'Content-type: text/html\n\nERROR: ack need duration\n'
+
+    envi["QUERY_STRING"] = 'hostname=toto&SERVICE=test&action=ack&cause=test&duration=10y'
+    ret = run_cgi(cgibin, UNIXSOCK, envi, True, maxclient=2)
+    print(f"RET={ret}")
+    assert "out" in ret
+    assert ret["out"] == b'Content-type: text/html\n\nSENDSTRING\n'
+
+    envi["QUERY_STRING"] = 'hostname=toto&SERVICE=test&action=disable'
+    ret = run_cgi(cgibin, UNIXSOCK, envi, False)
+    print(f"RET={ret}")
+    assert "out" in ret
+    assert ret["out"] == b'Content-type: text/html\n\nERROR: disable need cause\n'
+
+    envi["QUERY_STRING"] = 'hostname=toto&SERVICE=test&action=disable&cause=test'
+    ret = run_cgi(cgibin, UNIXSOCK, envi, False)
+    print(f"RET={ret}")
+    assert "out" in ret
+    assert ret["out"] == b'Content-type: text/html\n\nERROR: disable need duration\n'
+
+    envi["QUERY_STRING"] = 'hostname=toto&SERVICE=test&action=disable&cause=test&duration=10y'
+    ret = run_cgi(cgibin, UNIXSOCK, envi, True, maxclient=2)
+    print(f"RET={ret}")
+    assert "out" in ret
+    assert ret["out"] == b'Content-type: text/html\n\nERROR: disable need dsvc\n'
+
+    envi["QUERY_STRING"] = 'hostname=toto&service=test&action=disable&cause=test&duration=10y&dservice=toto'
+    ret = run_cgi(cgibin, UNIXSOCK, envi, True, maxclient=2)
+    print(f"RET={ret}")
+    assert "out" in ret
+    assert ret["out"] == b'Content-type: text/html\n\nSENDSTRING\n'
+
+    envi["QUERY_STRING"] = 'HOST=toto&SERVICE=test&action=disable&cause=test&duration=10y&DSERVICE=toto'
+    ret = run_cgi(cgibin, UNIXSOCK, envi, True, maxclient=2)
+    print(f"RET={ret}")
+    assert "out" in ret
+    assert ret["out"] == b'Content-type: text/html\n\nSENDSTRING\n'
+
+    envi["QUERY_STRING"] = 'HOST=toto&SERVICE=test&TIMEBUF=tutu'
+    ret = run_cgi(cgibin, UNIXSOCK, envi, False)
+    print(f"RET={ret}")
+    assert "out" in ret
+    assert "timeout" not in ret
+    assert ret["recv"] == b'GETSTATUS toto test tutu\n'
+    assert ret["out"] == b'Content-type: text/html\n\nSENDSTRING\n'
+
+    us = socket.socket(socket.AF_UNIX, socket.SOCK_STREAM)
+    us.bind(UNIXSOCK)
+
+    envi["QUERY_STRING"] = 'hostname=toto&SERVICE=test'
+    ret = run_cgi(cgibin, UNIXSOCK, envi, False)
+    print(f"RET={ret}")
+    assert "out" in ret
+    assert ret["recv"] == b'GETSTATUS toto test\n'
+    assert ret["out"] == b'Content-type: text/html\n\nSENDSTRING\n'
+
